+++ core/execution/market.go
@@ -1966,11 +1966,13 @@ func (m *Market) resolveClosedOutParties(ctx context.Context, distressedMarginEv
 	// we only need the MarketPosition events here, and rather than changing all the calls
 	// we can just keep the MarketPosition bit
 	closedMPs := make([]events.MarketPosition, 0, len(closed))
+	closedParties := make([]string, 0, len(closed))
 	// get the actual position, so we can work out what the total position of the market is going to be
 	var networkPos int64
 	for _, pos := range closed {
 		networkPos += pos.Size()
 		closedMPs = append(closedMPs, pos)
+		closedParties = append(closedParties, pos.Party())
 	}
 	if networkPos == 0 {
 		m.log.Warn("Network positions is 0 after closing out parties, nothing more to do",
@@ -2016,7 +2018,13 @@ func (m *Market) resolveClosedOutParties(ctx context.Context, distressedMarginEv
 			logging.Order(no),
 			logging.Error(err))
 	}
-
+	// Whether we were able to uncross the network order or not, we have to update the positions
+	// any closedParties element that wasn't previously marked as distressed should be marked as now
+	// being distressed. Any previously distressed positions that are no longer distressed
+	// should also be updated.
+	// If the network order uncrosses, we can ignore the distressed parties (they are closed out)
+	// but the safe parties should still be sent out.
+	dp, sp := m.position.MarkDistressed(closedParties)
 	// FIXME(j): this is a temporary measure for the case where we do not have enough orders
 	// in the book to 0 out the positions.
 	// in this case we will just return now, cutting off the position resolution
@@ -2024,7 +2032,19 @@ func (m *Market) resolveClosedOutParties(ctx context.Context, distressedMarginEv
 	// then when a new order is placed, the distressed parties will go again through positions resolution
 	// and if the volume of the book is acceptable, we will then process positions resolutions
 	if no.Remaining == no.Size {
+		// it is possible that we pass through here with the same distressed parties as before, no need
+		// to send the event if both distressed and safe parties slices are nil
+		if len(dp) != 0 || len(sp) != 0 {
+			devt := events.NewDistressedPositionsEvent(ctx, m.GetID(), dp, sp)
+			m.broker.Send(devt)
+		}
 		return orderUpdates, ErrNotEnoughVolumeToZeroOutNetworkOrder
+	}
+	// if we have any distressed positions that now no longer are distressed, emit the event
+	// no point in sending an event unless there's data
+	if len(sp) > 0 {
+		devt := events.NewDistressedPositionsEvent(ctx, m.GetID(), nil, sp)
+		m.broker.Send(devt)
 	}
 
 	// @NOTE: At this point, the network order was updated by the orderbook