+++ datanode/networkhistory/service_test.go
@@ -379,12 +379,12 @@ func TestMain(t *testing.M) {
 		log.Infof("%s", goldenSourceHistorySegment[4000].HistorySegmentID)
 		log.Infof("%s", goldenSourceHistorySegment[5000].HistorySegmentID)
 
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmVCYKHRBFmUm5Y7KcFg3meFvZgvud6o7ff2JoKoi6a59o", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmNjwTeX9EcEhHLh4DJdN6CxFyreWDbhMYf9YFzCgcm2TU", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmdTuZwcK74T3v2k87WzMEVUyjbae4wbKFzKrx5WVxht24", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmUVBq11sFMCQeJXLJt5mu3RzoeFiJw4W9DZbNoCdtBASh", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmQH1VWqNHvFW7X7eHZUFsxc9VgoLoscab288Ki1W7XtdY", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmWkqRUE1Y644CUm8YhX7S6XDa4nPCqYpNKrbSd6GxWieb", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmSxN7pFLSsaGXgJLzjyxBWQ4tF6uGLx6gxJ1KT8UxACaZ", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmTxDcB93HTfrnRg8e9ehyrxK9rf148UG6P32FZxNS55A6", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmbaY9KdWKSH4SSvJVir3LJoi4fLGfgRgeQLJFesoaHePj", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmU4REyMm4S6mLSvLFvhzdBeqyVo5K98n7QDBD2aDPMKEN", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmRoGYPZ9wmjzSW7VYsRVpfkguRjBM62SxxxfGXg2CXSbd", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmcbgKeEa3b7in2R1UH1tK429KE9TX1HsrMsbaGpSubGye", snapshots)
 	}, postgresRuntimePath, sqlFs)
 
 	if exitCode != 0 {