+++ datanode/sqlstore/migrations/0014_position_status.sql
@@ -1,0 +1,7 @@
+-- +goose Up
+
+ALTER TYPE position_status_type ADD VALUE IF NOT EXISTS 'POSITION_STATUS_DISTRESSED';
+
+-- +goose Down
+
+-- noop